refactor: GameMachine 패키지 이동
