+++ src/main/kotlin/wordle/Application.kt
@@ -1,5 +1,7 @@
 package wordle
 
+import wordle.service.GameMachine
+
 class Application {
 
     companion object {
